Render3D: - Add the ability to set the size of the output framebuffer. - Do some more code cleanup.

+++ desmume/src/texcache.h
@@ -73,7 +73,7 @@ public:
 	u32 sizeX, sizeY;
 	float invSizeX, invSizeY;
 
-	u64 texid; //used by ogl renderer for the texid
+	u32 texid; //used by ogl renderer for the texid
 	TexCache_TexFormat cacheFormat;
 
 	struct Dump {
